src/main/java/nextstep/subway/map/ui/MapController.java
package nextstep.subway.map.ui;

import nextstep.subway.map.application.MapService;
import nextstep.subway.map.dto.MapResponse;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/maps")
public class MapController {

    private final MapService mapService;

    public MapController(MapService mapService) {
        this.mapService = mapService;
    }

    @GetMapping
    public ResponseEntity<MapResponse> getMaps() {
        try {
            return ResponseEntity.ok(mapService.getMaps());
        } catch (RuntimeException e) {
            return ResponseEntity.badRequest().build();
        }
    }
}
